+++ SSAReplacement.Api/Domain/Jobs/JobVariable.cs
@@ -2,8 +2,8 @@ namespace SSAReplacement.Api.Domain;
 
 public class JobVariable
 {
-    public int Id { get; set; }
-    public int JobId { get; set; }
+    public long Id { get; set; }
+    public long JobId { get; set; }
     public required string Key { get; set; }
     public required string Value { get; set; }
